# perf(ext/http): Reduce size of `ResponseBytesInner`