@bartlomieju 2024-08-01
I see 8-9% speed up on M1 Max:
`main`

```
wrk -d 10s --latency http://localhost:8000
Running 10s test @ http://localhost:8000
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    75.96us   68.31us   3.40ms   99.36%
    Req/Sec    60.72k     2.53k   62.38k    97.03%
  Latency Distribution
     50%   73.00us
     75%   84.00us
     90%   97.00us
     99%  131.00us
  1220434 requests in 10.10s, 174.58MB read
Requests/sec: 120836.25
Transfer/sec:     17.29MB
```

This PR:
```
wrk -d 10s --latency http://localhost:8000
Running 10s test @ http://localhost:8000
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    67.71us   34.28us   3.12ms   92.86%
    Req/Sec    65.45k     1.55k   67.33k    93.07%
  Latency Distribution
     50%   68.00us
     75%   77.00us
     90%   89.00us
     99%  115.00us
  1315388 requests in 10.10s, 188.17MB read
Requests/sec: 130242.23
Transfer/sec:     18.63MB
```